+++ src/App.js
@@ -1,18 +1,17 @@
 import "./App.css";
 import Header from "./pages/Header";
 import { Routes, Route, Router } from "react-router-dom";
-import Home1 from "./layouts/Home1";
+import Home1 from "./pages/Home1";
 import CallToAction from "./pages/CallToAction";
-import Home2 from "./pages/Home2"
-import Lay05 from "./layouts/Lay05";
+import Main from "./layouts/Main";
+import Home2 from "./components/Home2";
 
 function App() {
   return (
     <>
-      <Lay05></Lay05>
+      <Main></Main>
     </>
   );
 }
 
 export default App;
-sprint-1
