+++ src/components/Home2.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import Logo from './Logo'
+import InicioU from './InicioU'
+import Empresa from './Empresa'
+import Carrusel from './Carrusel'
+import Footer from './Footer'
+
+export default function Home2() {
+
+  return (
+    <div>
+      <div className='cuerpo-general'>
+        <Carrusel/>
+      </div>
+    </div>
+  )
+}
