--- src/layouts/Lay05.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Header from "../pages/Header";
-import Footer from "../components/Footer"
-import Scrolltop from "../pages/Scrolltop";
-import { BrowserRouter as Router } from "react-router-dom";
-
-export default function Lay05(){
-    return (
-<>
-          <Header></Header>
-          <Footer></Footer>
-</>
-    );
-}
