+++ src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Home1 from "./Home1";
+import Home2 from "../components/Home2";
+
+export default function Home() {
+    return(
+        <>
+        <Home1></Home1>
+        <Home2></Home2>
+        </>
+    )
+}
